requirements.txt
groq
PyPDF2
spacy
streamlit
json
python-docx